requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
uvicorn==0.24.0
openai==1.3.7
httpx==0.25.2